toml: improved lexer/parser
We now use start conditions in the lexer. This makes everything much more
complicated, but it should make the planned metadata support easier.
For now, we mostly use start conditions for parsing strings and
distinguishing between BARE_STRING and FLOAT tokens.
This also changes how comments are parsed and stored.